build-failure-analyzer/pyproject.toml
[project]
name = "build-failure-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.2.1",
    "fastapi>=0.121.0",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-ollama>=1.0.0",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "slack-sdk>=3.37.0",
    "uvicorn[standard]>=0.38.0",
]
